requirements.txt
streamlit
pandas
plotly
prophet
openai
numpy
openpyxl
python-dotenv
